Add Neural Network Training
created a Jupyter notebook for training the galactic search neural
network, along with a self-contained python virtual environment for
executing it.

Moved the previous neural network code from Main.java into it's own
class.

Enhanced the build.gradle file so to add a deployment target, so you
don't need to use the web page to push new versions of the application.